+++ src/adapt/gro-adapter-node-library.ts
@@ -5,8 +5,8 @@ import {printSpawnResult, spawnProcess} from '../utils/process.js';
 import {TaskError} from '../task/task.js';
 import {copyDist} from '../build/dist.js';
 import {DIST_DIRNAME, paths, toBuildOutPath, toDistOutDir, toImportId} from '../paths.js';
-import {NODE_LIBRARY_BUILD_NAME} from '../config/defaultBuildConfig.js';
-import {BuildConfig, BuildName, printBuildConfigLabel} from '../config/buildConfig.js';
+import {NODE_LIBRARY_BUILD_NAME} from '../build/defaultBuildConfig.js';
+import {BuildConfig, BuildName, printBuildConfigLabel} from '../build/buildConfig.js';
 import {EMPTY_OBJECT} from '../utils/object.js';
 import {UnreachableError} from '../utils/error.js';
 import {stripEnd} from '../utils/string.js';
