+++ src/config/config.ts
@@ -1,7 +1,11 @@
 import {paths, groPaths, toBuildOutPath, CONFIG_BUILD_PATH, toImportId} from '../paths.js';
-import {isPrimaryBuildConfig, normalizeBuildConfigs, validateBuildConfigs} from './buildConfig.js';
+import {
+	isPrimaryBuildConfig,
+	normalizeBuildConfigs,
+	validateBuildConfigs,
+} from '../build/buildConfig.js';
 import type {AdaptBuilds} from '../adapt/adapter.js';
-import type {BuildConfig, BuildConfigPartial} from './buildConfig.js';
+import type {BuildConfig, BuildConfigPartial} from '../build/buildConfig.js';
 import {
 	LogLevel,
 	SystemLogger,
@@ -10,7 +14,10 @@ import {
 	DEFAULT_LOG_LEVEL,
 } from '../utils/log.js';
 import type {Logger} from '../utils/log.js';
-import {PRIMARY_NODE_BUILD_CONFIG, DEFAULT_ECMA_SCRIPT_TARGET} from './defaultBuildConfig.js';
+import {
+	PRIMARY_NODE_BUILD_CONFIG,
+	DEFAULT_ECMA_SCRIPT_TARGET,
+} from '../build/defaultBuildConfig.js';
 import type {EcmaScriptTarget} from '../build/tsBuildHelpers.js';
 import {omitUndefined} from '../utils/object.js';
 import type {ServedDirPartial} from '../build/ServedDir.js';
