+++ src/server.task.ts
@@ -1,5 +1,5 @@
 import type {Task} from './task/task.js';
-import {toApiServerBuildPath} from './config/defaultBuildConfig.js';
+import {toApiServerBuildPath} from './build/defaultBuildConfig.js';
 import {spawn} from './utils/process.js';
 import type {SpawnedProcess} from './utils/process.js';
 import {red} from './utils/terminal.js';
