+++ src/start.task.ts
@@ -6,10 +6,10 @@ import {loadGroConfig} from './config/config.js';
 import {spawn} from './utils/process.js';
 import type {SpawnedProcess} from './utils/process.js';
 import {green} from './utils/terminal.js';
-import type {BuildConfig} from './config/buildConfig.js';
+import type {BuildConfig} from './build/buildConfig.js';
 import {printTimings} from './utils/print.js';
 import {resolveInputFiles} from './build/utils.js';
-import {hasApiServer, hasSvelteKitFrontend, toApiServerPort} from './config/defaultBuildConfig.js';
+import {hasApiServer, hasSvelteKitFrontend, toApiServerPort} from './build/defaultBuildConfig.js';
 import type {TaskArgs as ServeTaskArgs} from './serve.task.js';
 import {toSvelteKitBasePath} from './build/sveltekit.js';
 import {loadPackageJson} from './utils/packageJson.js';
