+++ src/task/invokeTask.ts
@@ -21,7 +21,7 @@ import {findModules, loadModules} from '../fs/modules.js';
 import {plural} from '../utils/string.js';
 import {loadTaskModule} from './taskModule.js';
 import {loadGroPackageJson} from '../utils/packageJson.js';
-import {PRIMARY_NODE_BUILD_NAME} from '../config/defaultBuildConfig.js';
+import {PRIMARY_NODE_BUILD_NAME} from '../build/defaultBuildConfig.js';
 import type {Filesystem} from '../fs/filesystem.js';
 
 /*
